open-ai-sdk/requirements.txt
neo4j==5.11.0
fastapi==0.95.2
uvicorn==0.20.0
python-dotenv==1.0.0
openai==1.0.0
requests==2.31.0
